Added the Dark theme and future code to alternate it
The dark theme is its own .CSS file and DarkModeToggle will handle the switching of the two files. I also set dark theme to or standard CSS for now but that shouldn't conflict. Also there are a few lines of bootstrap that I changed to help match the CSS theme of the website. Should be harmless as far as other functions go but if we go with older versions, then some of the buttons and bars wont match the CSS dark theme.